[Refactor ⚙️] StyledButton -> StyledProfileActionButton 네이밍 수정
